Add scripts for Finder.app.
Adds scripts for comparing the selected files with an external diff application and for opening a Terminal.app shell for the selection.
Adds scripts for selecting files by shell glob and for opening Terminal.app or  selection in TextMate.app.  Also includes three icons to make droplets (and thus Finder toolbar icons) with.